Adding in tanks

+++ TowerDefense.Interfaces/Bullet.cs
@@ -5,7 +5,7 @@ namespace TowerDefense.Business.Models
     public class Bullet : IBullet
     {
         public double Range { get; set; }
-        public double Damage { get; set; }
+        public int Damage { get; set; }
         public double ReloadTime
         {
             get { return Range * Damage; }
